requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic_settings
python-dotenv
google-genai
mcp[cli]